Add dataset
Dataset Description
The dataset for this competition (both train and test) was generated from a deep learning model trained on the Used Car Price Prediction Dataset. Feature distributions are close to, but not exactly the same, as the original. Feel free to use the original dataset as part of this competition, both to explore differences as well as to see whether incorporating the original in training improves model performance.

Files
train.csv - the training dataset; price is the continuous target
test.csv - the test dataset; your objective is to predict the value of price for each row
sample_submission.csv - a sample submission file in the correct format